fix RobotGame not being imported correctly
