+++ frontend/src/components/Hint.astro
@@ -1,8 +1,13 @@
 ---
+interface Props {
+  visible: boolean;
+}
+
+const { visible = false } = Astro.props;
 
 ---
 
-<x-hint>
+<x-hint {visible}>
   <button>
     <!-- carbon:idea -->
     <!-- prettier-ignore -->
@@ -22,8 +27,10 @@
     grid-template-columns: auto;
 
     margin-top: 0rem;
-    border: 1px;
+    border: 1px var(--sl-color-white);
     border-radius: 0.25em;
+
+    opacity: 0.7;
   }
   button {
     background-color: var(--sl-color-gray-6);
@@ -69,6 +76,7 @@
   }
   x-hint[visible] {
     background-color: var(--sl-color-gray-6);
+    opacity: 1;
   }
   x-hint div {
     display: none;
@@ -81,7 +89,7 @@
     color: var(--sl-color-white);
   }
   x-hint[visible] button svg {
-    color: #dfc838;
+    opacity: 1;
   }
 </style>
