+++ frontend/src/components/RobotGame/RobotGame.ts
@@ -1,47 +1,8 @@
-import type { CodeQuestion } from "src/validation/CodeQuestion";
 import { confetti } from "src/utils/confetti";
 import type { CodeBlock, ResetEvent } from "@components/CodeBlock/CodeBlock";
 import type { EvalResponse } from "@components/CodeBlock/evaluate";
 import { parenthesisCheck } from "../../validation/0-robot";
-
-export const Functions = {
-  BASIC: "BASIC",
-  LOOK_HORIZONTAL: "LOOK_HORIZONTAL",
-  LOOKAROUND: "LOOKAROUND",
-} as const;
-export type Functions = (typeof Functions)[keyof typeof Functions];
-
-export type BoardProps = {
-  rows?: number;
-  cols?: number;
-  /** Position where the robot starts. */
-  start: number;
-  /** Enemy positions. */
-  enemies: number[];
-  /** Minimum steps needed to solve the puzzle. */
-  steps: number;
-};
-
-export type RobotGameProps = {
-  rows?: number;
-  cols?: number;
-  boards: BoardProps[];
-  /** Abilities available to the programmer. */
-  functions?: Functions;
-  /** If the boards need to be solved with the minimum steps. */
-  solveWithMinimumSteps?: boolean;
-  winText?: string;
-  loseText?: string;
-} & CodeQuestion;
-
-type Board = BoardProps & {
-  htmlTable: HTMLTableElement;
-  robot: HTMLImageElement;
-  tableChanged: boolean;
-  numEnemies: number;
-  rows: number;
-  cols: number;
-};
+import { type BoardProps, Functions, type Board, type RobotGameProps } from "./RobotGameTypes";
 
 export class RobotGame extends HTMLElement {
   readonly codeblock: CodeBlock;
