+++ frontend/src/content/docs/en/first-steps/1-introduction.mdx
@@ -32,7 +32,7 @@ export const errorMsg =
 `Woops, something went wrong and the code does not compile!
 If you've mistakenly messed up the code, click the "Reset" button to return it back to its original state.`
 
-<Hint>
+<Hint visible="true">
   I'm a hint 💡  
   Click me when you are stuck with a problem, but try solving it for yourself first!
 </Hint>
