+++ frontend/src/validation/5-5-robot.ts
@@ -1,4 +1,4 @@
-import { Functions, type RobotGameProps } from "@components/RobotGame/RobotGame";
+import { Functions, type RobotGameProps } from "@components/RobotGame/RobotGameTypes";
 import { codeMess, replace } from "./CodeQuestion";
 import { createRegExp, exactly, maybe, word } from "magic-regexp";
 import { _, end, semicolon, start } from "./regex";
